VMS-32453: Metadata SDK: Fix compiler warnings

+++ vms/server/plugins/analytics/stub_analytics_plugin/src/nx/vms_server_plugins/analytics/stub/object_detection/device_agent.cpp
@@ -75,7 +75,7 @@ Ptr<IMetadataPacket> DeviceAgent::generateObjectMetadataPacket(int64_t frameTime
         objects = generateObjects(kObjectAttributes, m_objectTypeIdsToGenerate);
     }
 
-    for (int i = 0; i < objects.size(); ++i)
+    for (int i = 0; i < (int) objects.size(); ++i)
     {
         objects[i]->setBoundingBox(generateBoundingBox(m_frameIndex, i, objects.size()));
         objects[i]->setTrackId(trackIdByTrackIndex(i));
